Add preliminary pipeline class structure and move functions to processing.
Not everything works, this is mostly a progress save
