+++ climateset/__init__.py
@@ -6,6 +6,7 @@ CONFIGS = PROJECT_ROOT / "configs"
 DATA_DIR = PROJECT_ROOT / "data"
 RAW_DATA = DATA_DIR / "raw"
 PROCESSED_DATA = DATA_DIR / "processed"
+FINAL_DATA = DATA_DIR / "final"
 LOAD_DATA = DATA_DIR / "load"
 META_DATA = DATA_DIR / "meta"
 SCRIPT_DIR = PROJECT_ROOT / "scripts"
